Fix and fixes.

+++ server/src/database/models.rs
@@ -2,16 +2,16 @@ use super::schema::ping_history;
 
 #[derive(Queryable)]
 pub struct PingHistoryItem {
-    pub id: i64,
+    pub id: i32,
     pub url: String,
-    pub iteration: i64,
-    pub ping: i64,
+    pub iteration: i32,
+    pub ping: i32,
 }
 
 #[derive(Insertable)]
 #[table_name = "ping_history"]
 pub struct NewPingHistoryItem<'a> {
     pub url : &'a str,
-    pub iteration: &'a i64 ,
-    pub ping: &'a i64,
+    pub iteration: &'a i32,
+    pub ping: &'a i32,
 }
